Created Password reset APIs

+++ src/main/java/com/memoize/api/Service/AuthServiceImpl.java
@@ -3,18 +3,24 @@ package com.memoize.api.Service;
 import com.memoize.api.Config.Common;
 import com.memoize.api.Dto.AuthenticationResponse;
 import com.memoize.api.Dto.LoginRequest;
+import com.memoize.api.Dto.PasswordResetRequest;
+import com.memoize.api.Dto.VerificationCodeCheckRequest;
 import com.memoize.api.Dto.SignupRequest;
 import com.memoize.api.Entity.User;
 import com.memoize.api.Entity.VerificationToken;
+import com.memoize.api.Enum.AuthSource;
 import com.memoize.api.Enum.Role;
+import com.memoize.api.Enum.VerificationType;
 import com.memoize.api.Repository.UserRepository;
 import com.memoize.api.Repository.VerificationTokenRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -43,7 +49,7 @@ public class AuthServiceImpl implements AuthService {
     @Override
     @Transactional
     public AuthenticationResponse signup(SignupRequest request) {
-        if (!verifyCode(request.email(), request.verificationCode())) {
+        if (!verifyCode(request.email(), request.verificationCode(), false)) {
             throw new BadCredentialsException("Invalid verification code");
         }
         User user = User.builder()
@@ -53,7 +59,7 @@ public class AuthServiceImpl implements AuthService {
                 .password(Common.PASSWORD_ENCODER.encode(request.password()))
                 .role(Role.USER)
                 .build();
-        userRepository.saveAndFlush(user);
+        userRepository.save(user);
         String jwtToken = jwtService.generateToken(user.getId().toString(), user.getRole().name());
         return AuthenticationResponse.of(jwtToken, user.getId());
     }
@@ -70,25 +76,50 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     @Transactional
-    public void sendVerificationEmail(String email, boolean newRegistration) {
-        String verificationCode = Common.generateRandomString(6);
-        if (newRegistration && !isEmailAvailable(email)) {
+    public void sendVerificationEmail(String email, VerificationType verificationType) {
+        Boolean isEmailAvailable = isEmailAvailable(email);
+        if (verificationType == VerificationType.VERIFY_EMAIL && !isEmailAvailable) {
             throw new IllegalArgumentException("Email already in use");
         }
-        if (!newRegistration && isEmailAvailable(email)) {
+        if (verificationType == VerificationType.RESET_PASSWORD && isEmailAvailable) {
             throw new IllegalArgumentException("Email not found");
         }
+
         if (verificationTokenRepository.existsValidTokenByEmail(email)) {
             return;
         }
+
+        String verificationCode = Common.generateRandomString(6);
+
         var verificationToken = VerificationToken.builder()
                 .token(Common.encodeBase64(verificationCode)).email(email).build();
-        verificationTokenRepository.saveAndFlush(verificationToken);
-        mailService.sendVerificationEmail(email, verificationCode);
+        verificationTokenRepository.save(verificationToken);
+        mailService.sendVerificationCodeEmail(email, verificationCode, verificationType);
+    }
+
+    @Override
+    @Transactional
+    public void verifyPasswordResetCode(VerificationCodeCheckRequest request) {
+        if (!verifyCode(request.email(), request.verificationCode(), true)) {
+            throw new BadCredentialsException("Invalid verification code");
+        }
+    }
+
+    @Override
+    @Transactional
+    public void resetPassword(PasswordResetRequest request) {
+        if (!verifyCode(request.email(), request.verificationCode(), false)) {
+            throw new BadCredentialsException("Invalid verification code");
+        }
+        User currentUser = userRepository.findByEmail(request.email())
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+        currentUser.setPassword(Common.PASSWORD_ENCODER.encode(request.newPassword()));
+        userRepository.save(currentUser);
     }
 
     // helper
-    private boolean verifyCode(String email, String code) {
+    @Transactional
+    protected boolean verifyCode(String email, String code, boolean toCheckAgain) {
         String encodedCode = Common.encodeBase64(code);
         var token = verificationTokenRepository.findByEmailAndToken(email, encodedCode);
         if (token.isEmpty()) {
@@ -98,7 +129,9 @@ public class AuthServiceImpl implements AuthService {
             verificationTokenRepository.delete(token.get());
             return false;
         }
-        verificationTokenRepository.delete(token.get());
+        if (!toCheckAgain) {
+            verificationTokenRepository.delete(token.get());
+        }
         return true;
     }
 }
